+++ src/main/java/hanium/modic/backend/domain/user/service/UserVoteStreakService.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Service;
 
 import hanium.modic.backend.common.error.exception.LockException;
 import hanium.modic.backend.common.property.property.VoteProperties;
-import hanium.modic.backend.common.redis.distributedLock.LockManager;
+import hanium.modic.backend.infra.redis.distributedLock.LockManager;
 import hanium.modic.backend.domain.user.entity.UserVoteStreak;
 import hanium.modic.backend.domain.user.repository.UserVoteStreakRepository;
 import lombok.RequiredArgsConstructor;
