+++ src/main/java/hanium/modic/backend/domain/vote/service/VoteCompletionRewardService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import hanium.modic.backend.common.error.exception.AppException;
-import hanium.modic.backend.common.redis.distributedLock.LockManager;
+import hanium.modic.backend.infra.redis.distributedLock.LockManager;
 import hanium.modic.backend.common.error.exception.LockException;
 import hanium.modic.backend.domain.ticket.service.TicketService;
 import hanium.modic.backend.domain.vote.entity.SimilarityVoteResultEntity;
