+++ src/main/java/hanium/modic/backend/domain/vote/service/VotingService.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import hanium.modic.backend.common.error.exception.AppException;
 import hanium.modic.backend.common.error.exception.LockException;
 import hanium.modic.backend.common.property.property.VoteProperties;
-import hanium.modic.backend.common.redis.distributedLock.LockManager;
+import hanium.modic.backend.infra.redis.distributedLock.LockManager;
 import hanium.modic.backend.domain.user.service.UserVoteStreakService;
 import hanium.modic.backend.domain.vote.dto.VoteRewardResult;
 import hanium.modic.backend.domain.vote.entity.SimilarityVoteEntity;
