+++ src/main/java/hanium/modic/backend/infra/redis/config/RedissonConfig.java
@@ -1,4 +1,4 @@
-package hanium.modic.backend.common.redis.config;
+package hanium.modic.backend.infra.redis.config;
 
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
