+++ src/main/java/hanium/modic/backend/infra/redis/distributedLock/LockManager.java
@@ -1,4 +1,4 @@
-package hanium.modic.backend.common.redis.distributedLock;
+package hanium.modic.backend.infra.redis.distributedLock;
 
 import java.util.List;
 import java.util.concurrent.TimeUnit;
