+++ src/test/java/hanium/modic/backend/domain/vote/service/VotingServiceTest.java
@@ -11,7 +11,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import hanium.modic.backend.common.property.property.VoteProperties;
-import hanium.modic.backend.common.redis.distributedLock.LockManager;
+import hanium.modic.backend.infra.redis.distributedLock.LockManager;
 import hanium.modic.backend.domain.post.repository.PostEntityRepository;
 import hanium.modic.backend.domain.user.service.UserVoteStreakService;
 import hanium.modic.backend.domain.vote.repository.SimilarityVoteRepository;
